pyproject.toml
[tool.poetry]
name = "zugzwang-core"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-socketio = "^4.4"
aiohttp = "^3.6"
pytz = "^2019.3"
cards = "^0.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
